+++ tests/testdata/zimui/assets/index.js
@@ -1,0 +1,1 @@
+Fake index.js for testing.